implementing corrected colorscheme for TeX files
TeX files did not use the proper colors when they were loaded. This
commit changes some of the default colors and adds a full copy of
the (modified) deathstar colorscheme adapted for TeX. In the future,
this complete colorscheme should be pared down, since the only
adjustments were made to the 'Programming Languages' section (the
adjustments might also be made more TeX specific; check :highlight
to see the current TeX settings).